+++ scrapegraphai/models/__init__.py
@@ -11,3 +11,5 @@ from .ollama import Ollama
 from .hugging_face import HuggingFace
 from .groq import Groq
 from .bedrock import Bedrock
+from .anthropic import Anthropic
+from .claude import Claude