+++ scrapegraphai/nodes/fetch_node.py
@@ -1,10 +1,12 @@
 """ 
 FetchNode Module
 """
-
+import pandas as pd
+import json
 from typing import List, Optional
 from langchain_community.document_loaders import AsyncChromiumLoader
 from langchain_core.documents import Document
+from langchain_community.document_loaders import PyPDFLoader
 from .base_node import BaseNode
 from ..utils.remover import remover
 
@@ -21,7 +23,7 @@ class FetchNode(BaseNode):
     Attributes:
         headless (bool): A flag indicating whether the browser should run in headless mode.
         verbose (bool): A flag indicating whether to print verbose output during execution.
-    
+
     Args:
         input (str): Boolean expression defining the input keys needed from the state.
         output (List[str]): List of output keys to be updated in the state.
@@ -29,11 +31,13 @@ class FetchNode(BaseNode):
         node_name (str): The unique identifier name for the node, defaulting to "Fetch".
     """
 
-    def __init__(self, input: str, output: List[str], node_config: Optional[dict], node_name: str = "Fetch"):
+    def __init__(self, input: str, output: List[str], node_config: Optional[dict] = None, node_name: str = "Fetch"):
         super().__init__(node_name, "node", input, output, 1)
 
-        self.headless = True if node_config is None else node_config.get("headless", True)
-        self.verbose = True if node_config is None else node_config.get("verbose", False)
+        self.headless = True if node_config is None else node_config.get(
+            "headless", True)
+        self.verbose = False if node_config is None else node_config.get(
+            "verbose", False)
 
     def execute(self, state):
         """
@@ -56,16 +60,39 @@ class FetchNode(BaseNode):
 
         # Interpret input keys based on the provided input expression
         input_keys = self.get_input_keys(state)
-
         # Fetching data from the state based on the input keys
         input_data = [state[key] for key in input_keys]
 
         source = input_data[0]
-        if self.input == "json_dir" or self.input == "xml_dir":
+        if self.input == "json_dir" or self.input == "xml_dir" or self.input == "csv_dir":
             compressed_document = [Document(page_content=source, metadata={
                 "source": "local_dir"
             })]
         # if it is a local directory
+
+        # handling for pdf
+        elif self.input == "pdf":
+            loader = PyPDFLoader(source)
+            compressed_document = loader.load()
+
+        elif self.input == "csv":
+            compressed_document = [Document(page_content=str(pd.read_csv(source)), metadata={
+                "source": "csv"
+            })]
+        elif self.input == "json":
+            f = open(source)
+            compressed_document = [Document(page_content=str(json.load(f)), metadata={
+                "source": "json"
+            })]
+        elif self.input == "xml":
+            with open(source, 'r', encoding='utf-8') as f:
+                data = f.read()
+            compressed_document = [Document(page_content=data, metadata={
+                "source": "xml"
+            })]
+        elif self.input == "pdf_dir":
+            pass
+
         elif not source.startswith("http"):
             compressed_document = [Document(page_content=remover(source), metadata={
                 "source": "local_dir"
@@ -73,7 +100,7 @@ class FetchNode(BaseNode):
 
         else:
             if self.node_config is not None and self.node_config.get("endpoint") is not None:
-                
+
                 loader = AsyncChromiumLoader(
                     [source],
                     proxies={"http": self.node_config["endpoint"]},